requirements.txt
stockfish
python-chess
pygame~=1.9.6